+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Update Rust and Go libs ([#46](https://github.com/stackabletech/stackablectl/pull/46))
+- Support generation of shell completions ([#54](https://github.com/stackabletech/stackablectl/pull/54))
 
 ## [0.2.0] - 2022-05-17